requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
playwright==1.58.0
google-generativeai==0.8.6
python-dotenv==1.2.2
pydantic==2.12.5
requests==2.33.1
